requirements.txt
fastapi
uvicorn
langchain
openai
pypdf
chromadb
tiktoken
python-dotenv
